cleanup(js): move createGlobPatternsForDependencies to js
