+++ packages/js/src/utils/generate-globs.ts
